Conflicts:
	MultiBrowser2/BrowserFrame.cpp
	MultiBrowser2/BrowserWidget.hpp
	MultiBrowser2/EmbedWidget.hpp
	MultiTouch/BinaryServer.cpp
	MultiTouch/Merge.hpp
	MultiWidgets/Animator.hpp
	MultiWidgets/Animators.hpp
	MultiWidgets/CloudWidget.cpp
	MultiWidgets/GridContainer.hpp
	MultiWidgets/ListWidget.cpp
	MultiWidgets/TextBox.cpp
	MultiWidgets/Widget.cpp
	Plugins/Draw/Draw.cpp
	Plugins/Draw/Draw.hpp
	Plugins/Draw/Stroke.cpp
	Plugins/TimeLinePlugin/Circles.cpp
	multitude/Radiant/BinaryData.cpp
	multitude/Radiant/DirectoryCommon.cpp
	multitude/Radiant/DirectoryQt.cpp
	multitude/Radiant/FileUtils.cpp
	multitude/Radiant/Trace.cpp
	multitude/Valuable/HasValues.cpp
	multitude/multitude.pri

+++ multitude/Luminous/Texture.hpp
@@ -157,9 +157,9 @@ namespace Luminous
     virtual long consumesBytes()
     {
       /// @todo how about compressed formats with mipmaps, does the 4/3 rule apply here as well?
-      if(m_consumed > 0) return m_consumed;
+      if(m_consumed > 0) return (long)m_consumed;
 
-      return estimateMemoryUse();
+      return (long)estimateMemoryUse();
     }
 
     /// Get the OpenGL texture id
